chore: minor bugfixes and readme updated
Solved minor bugfixes in case malloc fails, change data types to be more consistent accross platforms: char to int in ak_strchr() to prevent bugs for char > 127
Changed int to ssize_t to be more consistent in read_to_newline()

+++ get_next_line_bonus.c
@@ -6,7 +6,7 @@
 /*   By: akdovlet <akdovlet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/13 22:42:25 by akdovlet          #+#    #+#             */
-/*   Updated: 2024/01/13 22:42:28 by akdovlet         ###   ########.fr       */
+/*   Updated: 2026/05/20 08:50:53 by akdovlet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*read_to_newline(int fd, char *buffer)
 {
-	char	*tmp;
-	int		i;
+	char		*tmp;
+	ssize_t		i;
 
 	tmp = NULL;
 	if (buffer[0])
